requirements.txt
# 大模型API客户端
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0

# 图像处理
pillow>=9.0.0

# 实用工具
python-dotenv>=1.0.0
requests>=2.28.0 